FIX. killing app through UI doesn't kill actual app instance at the sender
if the streaming thread is finished. sail::swapBuffer() will block permanently.
(because of semaphore operation on the double buffer b/w sail::swapBuffer() and SBS::streamLoop() )
If it blocks permanently, app could never get sageMessage which causes the killing app instance issue.
So, if the streaming thread is finished, then swapBuffer() should return with error so that APP_QUIT message can be DELIVERED.

-Sungwon